Modifications

+++ AccountCreationApp/models.py
@@ -7,8 +7,8 @@ class CustomerInfo(AbstractUser):
 	
 	Age = models.IntegerField(default=0)
 	Address= models.CharField(max_length=20,default=0)
-	AccountNumber=models.IntegerField(unique=True,default=0)
-	Balance= models.IntegerField(default=0)
+	AccountNumber=models.BigIntegerField(unique=True,default=0)
+	Balance= models.BigIntegerField(default=0)
 	Q1=models.CharField(max_length=20,default=0)
 	Q2=models.CharField(max_length=20,default=0)
 	Q3=models.CharField(max_length=20,default=0)
